improve performance for orderbook

+++ src/models/basic_types.hpp
@@ -5,7 +5,7 @@
 #include <string>
 
 namespace stockex::models {
-using OrderId = uint64_t;
+using OrderId = uint32_t;
 constexpr auto INVALID_ORDER_ID = std::numeric_limits<OrderId>::max();
 inline auto orderIdToString(OrderId id) -> std::string {
   if (id == INVALID_ORDER_ID) {
@@ -25,7 +25,7 @@ inline auto instrumentIdToString(InstrumentId id) -> std::string {
   return std::to_string(id);
 }
 
-using ClientId = uint32_t;
+using ClientId = uint8_t;
 constexpr auto INVALID_CLIENT_ID = std::numeric_limits<ClientId>::max();
 inline auto clientIdToString(ClientId id) -> std::string {
   if (id == INVALID_CLIENT_ID) {
@@ -35,7 +35,7 @@ inline auto clientIdToString(ClientId id) -> std::string {
   return std::to_string(id);
 }
 
-using Price = int64_t;
+using Price = uint16_t;
 constexpr auto INVALID_PRICE = std::numeric_limits<Price>::max();
 inline auto priceToString(Price price) -> std::string {
   if (price == INVALID_PRICE) {
@@ -45,7 +45,7 @@ inline auto priceToString(Price price) -> std::string {
   return std::to_string(price);
 }
 
-using Quantity = uint32_t;
+using Quantity = uint16_t;
 constexpr auto INVALID_QUANTITY = std::numeric_limits<Quantity>::max();
 inline auto quantityToString(Quantity quantity) -> std::string {
   if (quantity == INVALID_QUANTITY) {
